backend/requirements.txt
# Core dependencies
fastapi>=0.110.1
uvicorn>=0.25.0
python-dotenv>=1.0.1

# Database
sqlalchemy>=2.0.23
alembic>=1.10.2
psycopg2-binary>=2.9.0
pymongo>=4.5.0
motor>=3.3.1

# Data processing
pandas>=2.2.0
numpy>=1.26.0
python-dateutil>=2.8.2
pytz>=2023.3

# API and Security
requests>=2.31.0
python-jose[cryptography]>=3.3.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
python-multipart>=0.0.9
requests-oauthlib>=2.0.0
cryptography>=42.0.8

# AWS and other services
boto3>=1.34.129

# Development and testing
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
jq>=1.6.0
typer>=0.9.0